Singleton KahunaAppLaunchMonitor name changed to KahunaPushMonitor. Added properties pushInfo, applicationState & kahunaInitialized which will be used to store any push message, the application state and if kahuna has been initialized or not. We now swizzle the following methods when the application launches : - (void) application:(UIApplication *)application didFailToRegisterForRemoteNotificationsWithError:(NSError *)error - (void) application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo - (void) application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo fetchCompletionHandler:(void (^) (UIBackgroundFetchResult result))completionHandler - (void)application:(UIApplication *)application handleActionWithIdentifier:(NSString *)identifier forRemoteNotification:(NSDictionary *)userInfo completionHandler:(void(^)())completionHandler
The method "- (void) application:(UIApplication *)application didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken" is not being swizzled since we already have a API "registerForRemoteNotificationsWithDeviceToken" that Segment will call to send us the push token.

While the above swizzling is also possible inside the Kahuna SDK, we do not want to make any Kahuna calls till the "start" method is called. Also we want to make sure we swizzle on the main thread and not on a background thread. That is why we swizzle when we get the applicationDidFinishLaunching notification.
When we get remote notifications we check the value of "kahunaInitialized" on KahunaPushMonitor. If kahuna is initialized we call handleNotification API, otherwise we store it for future use. This is also useful when we have timing issues between a push coming in on App launch and the "Start" method has not been called yet. In those cases we check the value of "pushInfo" inside the "start" method and  then decide based on that whether to call handleNotification API or not.