Improve Apptimize Root integration
Previously everytime the Apptimize integration would be loaded, we would
fire an event for every experiment viewed by the user.

(It didn't actually work since Apptimize only returned experimements for
the session, which would almost alwyas be zero).

Now we hook into Apptimize to detect when an A/B test is viewed, and
fire the event then.