Store `userId` and `anonymousId` in NSUserDefaults
When looking for the `userId` and `anonymousId`, look first in
`NSUserDefaults`, then in the stored file. The stored file is persistent
in most cases, but on tvOS, the file system should be treated as
temporary. That means we lose the identity of the user is lost as soon
as the file system is cleared on the Apple TV.

This solution does not replace these file, but simply augments them,
using `NSUserDefaults` as a cache in front of the files. The only
rationale I can think of for using files is that sometimes apps blow
away their defaults.

There is probably another approach here which finds a happy medium, but
`NSUserDefaults` is all we've really got (unless you want to play with
the keychain, and I sure don't) on tvOS.