Record entire context for every event.
Previously, some pieces of context were set on the batch (static
context such as library info, device info, etc.).

This had the advantage of reducing the payload size (send it once
instead of 20 times in 20 event batch). But by adding gzip support we
don't really need to do this as much anymore.

So now we store the entire context with every payload, which represents the
context at the time of the event (rather than upload) more accurately.