Ensure network.carrier is populated.
Previously liveContext would populate `network.cellular` and
`network.wifi`, whilc staticContext would populate `network.carrier`.

In merging the two, the `network` dictionary in liveContext would
override the one in staticContext, causing `network.carrier` to be
ignored in the merged dictionary.

For now, we're populating the `network.carrier` as part of liveContext
to avoid this conflict.

Another potential fix would be to do a deep merge instead of shallow merge so
that `network.carrier` is preserved. However this would also apply to
custom context dictionaries sent by customers, and might not be the
behaviour they want (e.g. if they sent `context.network` as `{}`, the
deep merge would preserve the fields we collect, while a shallow merge
would ignore it).