fix: transaction.transactionIdentifier is nullable
According to Apple’s documentation (https://developer.apple.com/reference/storekit/skpaymenttransaction) the transactionIdentifier attribute can be nil.

Trying to create a dictionary with a nil value leads to a crash, and we’ve seen this type of crash in production (http://crashes.to/s/8c3d076840e). The nil value would be used with the keys “orderId” and “sku”.