Create HTTP requests on a background queue.
This refactors the internal implementation of the `sendData` to _create_
the HTTP requests on a new background queue. Note that this simply changes
the _creation_ of the request to the new background queue - the actual
execution was already on a background thread.

With this change, customers to add more complex logic in their connection
factories. e.g. a customer can use the connection factory to fetch a
token for a user, and attach that token to the request being made to by
the library to their proxy.

Previously, the creation would happen on the queueing thread. This
means that while the HTTP request was being _created_, events would not be
queued. By default, this is not a problem, the actual creation of the
request is a few ms at most. However, this assumes that the connection
factory customers might set only runs for a few ms. A long running connection
factory would have a chance of blocking the queueing thread for too long and
open a small window for data loss.

This opened up a slight chance that the integration would call flush
(for various reasons) while the HTTP request was being created
(particularly if the connection factory is slow). Hence we add a check to
verify that a request is not in progress before proceeding to create the
request. This is done with a boolean flag that is only read/updated on
the serialQueue.

Lastly, the code that sets the `sentAt` flag is now moved to inside the
HTTP Client. This is to move it closer to the time when the HTTP request
is actually made - otherwise a slow connection factory would cause the
server to think that the device timestamps are skewed.