Add End to End Tests (#746)
Sets up End to End Tests as described in https://paper.dropbox.com/doc/Libraries-End-to-End-Tests-ESEakc3LxFrqcHz69AmyN.

Currently our libraries only test with an in memory mock API. Using the real API alone isn’t great either as it always returns a 200. This is problematic because often our tests don’t mimic the validation done in the real pipeline.

A dedicated Segment source is setup with a webhook integration to Runscope. The test is configured with the write key of this source, and sends unique identifier with each event.

Runscope captures the requests it receives as the webhook integration. The E2E test uses this API to read all the events sent to Runscope, and verifies that our intended event made it to Runscope.

The screen tracking tests also rely on some global state that conflicts with these tests, so I've removed those tests for now - the end to end tests are more important.