Ensure queue is always < 1000 items.
Previously we didn't cap the queue, and when we added the limit, we only removed the oldest element, rather than fully shrink the queue.

Hence there are cases where the queue may already be larger than 1000 events. Now we delete as many events as required to trim the queue size.

Ref: LIB-210